refactoring Module and Parameter into high level abstractions file.
